testapp/requirements/dev.txt
-r common.txt
django-extensions==1.1.1
Werkzeug<1.0